Simplify to use style engine, rename to use-dimensions-props.js, drop __experimental prefix
