+++ packages/block-editor/src/index.js
@@ -6,7 +6,7 @@ export {
 	getBorderClassesAndStyles as __experimentalGetBorderClassesAndStyles,
 	useBorderProps as __experimentalUseBorderProps,
 	getColorClassesAndStyles as __experimentalGetColorClassesAndStyles,
-	getDimensionsClassesAndStyles as __experimentalGetDimensionsClassesAndStyles,
+	getDimensionsClassesAndStyles,
 	getTypographyClassesAndStyles,
 	useColorProps as __experimentalUseColorProps,
 	useCustomSides as __experimentalUseCustomSides,
